version 4.3.0 (VCL+FMX) - Updated to Image32 4.6 Released 26 Nov 2024 to fix issue drawing text inside SVG - NameSpace added to all uses - Added SVGIconImageListRestClient package - New Option in Component Editor to select Icons from WEB (VCL and FMX) - Demos updated

+++ Image32/source/Img32.Transform.pas
@@ -3,7 +3,7 @@ unit Img32.Transform;
 (*******************************************************************************
 * Author    :  Angus Johnson                                                   *
 * Version   :  4.6                                                             *
-* Date      :  18 September 2024                                               *
+* Date      :  26 November 2024                                                *
 * Website   :  http://www.angusj.com                                           *
 * Copyright :  Angus Johnson 2019-2024                                         *
 * Purpose   :  Affine and projective transformation routines for TImage32      *
@@ -1204,9 +1204,9 @@ begin
   begin
     a := a * Cardinal(w);
     inc(fAlphaTot, a);
-    inc(fColorTotB, (a * Byte(c)));
-    inc(fColorTotG, (a * Byte(c shr 8)));
-    inc(fColorTotR, (a * Byte(c shr 16)));
+    inc(fColorTotB, (Int64(a) * Byte(c)));
+    inc(fColorTotG, (Int64(a) * Byte(c shr 8)));
+    inc(fColorTotR, (Int64(a) * Byte(c shr 16)));
   end;
 end;
 //------------------------------------------------------------------------------
